requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.3.3
bs4==0.0.2
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
decorator==5.1.1
deepdiff==7.0.1
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.3
langchain-core==0.2.23
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.1.9
langsmith==0.1.77
linkify-it-py==2.0.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.35.14
ordered-set==4.1.0
orjson==3.10.4
packaging==23.2
pandas==2.1.4
panel==1.3.6
param==2.0.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
playwright==1.44.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.1.0
Pygments==2.18.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tenacity==8.3.0
tiktoken==0.7.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
types-requests==2.32.0.20240712
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
urllib3==2.2.2
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
